Improve flow support
