+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "debugger.html",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "license": "MPL-2.0",
   "repository": {
     "url": "git://github.com/devtools-html/debugger.html.git",
@@ -26,6 +26,7 @@
     "prepush": "npm run lint; node public/js/test/node-unit-tests.js --dots",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
+    "flow-coverage": "flow-coverage-report -i 'public/js/actions/*.js' -i 'public/js/reducers/*.js' -i 'public/js/utils/*.js' -t html -t text",
     "preinstall": "./bin/preinstall"
   },
   "dependencies": {
@@ -48,12 +49,25 @@
     "babel-cli": "^6.7.5",
     "babel-core": "^6.17.0",
     "babel-eslint": "^6.1.2",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-module-resolver": "^2.2.0",
+    "babel-plugin-transform-async-to-generator": "^6.16.0",
+    "babel-plugin-transform-es2015-block-scoping": "^6.7.1",
+    "babel-plugin-transform-es2015-destructuring": "^6.6.5",
+    "babel-plugin-transform-es2015-parameters": "^6.7.0",
+    "babel-plugin-transform-es2015-spread": "^6.6.5",
+    "babel-plugin-transform-flow-strip-types": "^6.14.0",
+    "babel-plugin-transform-runtime": "^6.7.5",
+    "babel-plugin-webpack-alias": "^2.1.1",
+    "babel-polyfill": "^6.7.4",
+    "babel-register": "^6.7.2",
     "body-parser": "^1.15.0",
     "check-node-version": "^1.1.2",
+    "classnames": "^2.2.5",
     "co": "=4.6.0",
     "css-loader": "^0.25.0",
     "devtools-config": "^0.0.3",
-    "devtools-local-toolbox": "^0.0.5",
+    "devtools-local-toolbox": "^0.0.6",
     "devtools-modules": "^0.0.3",
     "devtools-network-request": "^0.0.3",
     "devtools-sham-modules": "^0.0.3",
@@ -67,6 +81,7 @@
     "extract-text-webpack-plugin": "^1.0.1",
     "firefox-profile": "^0.4.0",
     "flow-bin": "^0.33.0",
+    "flow-coverage-report": "^0.1.0",
     "geckodriver": "^1.1.2",
     "glob": "^7.0.3",
     "husky": "^0.11.7",
@@ -78,44 +93,26 @@
     "mustache": "^2.2.1",
     "node-static": "^0.7.7",
     "npm": "^3.10.7",
+    "react": "=0.14.7",
+    "react-dom": "=0.14.7",
     "react-hot-loader": "^1.3.0",
+    "react-immutable-proptypes": "^1.7.1",
+    "react-inlinesvg": "^0.5.3",
+    "react-redux": "4.4.5",
+    "redux": "3.5.2",
     "rimraf": "^2.5.2",
     "selenium-webdriver": "^3.0.0-beta-2",
     "serve-index": "^1.8.0",
     "style-loader": "^0.13.1",
     "stylelint": "^7.4.2",
+    "svg-inline-loader": "^0.7.1",
+    "svg-inline-react": "^1.0.2",
+    "tcomb": "^3.1.0",
     "webpack": "1.13.1",
     "webpack-dev-middleware": "^1.6.1",
     "webpack-hot-middleware": "^2.12.0",
     "workerjs": "github:jlongster/workerjs",
-    "ws": "^1.0.1",
-    "babel-cli": "^6.7.5",
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-module-resolver": "^2.2.0",
-    "babel-plugin-transform-async-to-generator": "^6.16.0",
-    "babel-plugin-transform-es2015-block-scoping": "^6.7.1",
-    "babel-plugin-transform-es2015-destructuring": "^6.6.5",
-    "babel-plugin-transform-es2015-parameters": "^6.7.0",
-    "babel-plugin-transform-es2015-spread": "^6.6.5",
-    "babel-plugin-transform-flow-strip-types": "^6.14.0",
-    "babel-plugin-transform-runtime": "^6.7.5",
-    "babel-plugin-webpack-alias": "^2.1.1",
-    "babel-polyfill": "^6.7.4",
-    "babel-register": "^6.7.2",
-    "devtools-config": "^0.0.3",
-    "react": "=0.14.7",
-    "react-dom": "=0.14.7",
-    "react-immutable-proptypes": "^1.7.1",
-    "react-inlinesvg": "^0.5.3",
-    "react-redux": "4.4.5",
-    "redux": "3.5.2",
-    "classnames": "^2.2.5",
-    "tcomb": "^3.1.0",
-    "serve-index": "^1.8.0",
-    "svg-inline-loader": "^0.7.1",
-    "svg-inline-react": "^1.0.2"
+    "ws": "^1.0.1"
   },
   "files": [
     "public"
