+++ packages/devtools-local-toolbox/package.json
@@ -1,6 +1,6 @@
 {
   "name": "devtools-local-toolbox",
-  "version": "0.0.5",
+  "version": "0.0.6",
   "license": "MPL-2.0",
   "repository": {
     "type": "git",
