+++ packages/devtools-local-toolbox/public/js/clients/chrome/events.js
@@ -1,4 +1,4 @@
-const { Source, Location, Frame } = require("../../types");
+const { Source, Location, Frame } = require("../../tcomb-types");
 
 let actions;
 let pageAgent;
