+++ packages/devtools-local-toolbox/public/js/clients/firefox.js
@@ -2,7 +2,7 @@ const { DebuggerClient, DebuggerTransport,
         TargetFactory, WebsocketTransport } = require("devtools-sham-modules");
 const defer = require("../utils/defer");
 const { getValue } = require("devtools-config");
-const { Tab } = require("../types");
+const { Tab } = require("../tcomb-types");
 const { setupCommands, clientCommands } = require("./firefox/commands");
 const { setupEvents, clientEvents } = require("./firefox/events");
 const { createSource } = require("./firefox/create");
