--- public/js/flow-types.js
@@ -1,16 +1,0 @@
-/**
- * Flow types
- * @module types
- */
-
-/**
- * Source
- *
- * @memberof types
- * @static
- */
-export type Source = {
-  id: string,
-  url?: string,
-  sourceMapURL?: string
-};
